pyproject.toml
[tool.poetry]
name = "elite-qualifier"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
bs4 = "^0.0.1"
pandas = "^1.2.0"
praw = "^7.1.0"
pyspellchecker = "^0.5.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
